Support `VARIANCE` aggregate function (#598)
1. Add VARIANCE aggregation function
ref. https://www.ibm.com/docs/ko/db2/11.1?topic=functions-variance

VARIANCE = SUM(expression**2)/COUNT(expression) - (SUM(expression)/COUNT(expression))**2
The following variance basic formula causes the division to be performed twice, resulting in relative error. Which means, the error becomes larger than in EPSILON, so there is a problem in comparing decimal points in the test.
Therefore, the formula has been modified so that division is performed only once.

2. modified avg aggregate test
The test that compared the existing average value as an integer is modified to a test that compares the decimal result.