Fix `CAST` into i32 or i64 from numeric literals (#611)
select cast ( x as int(32)) from ... will output i32::MAX when x is greater than I32::MAX, and will put i32::MIN in for x for values less than i32::MIN. When x is not in i32 range, it should output an error.

This fix does cause one test to fail:
SELECT CAST(1.1 AS INTEGER) AS cast FROM Item"#,
should succeed but fails with this PR. Do we want this to get converted to an integer value (ie, truncated?) I believe mysql does truncate the float to an integer.

All solutions I have tried so far causes one of the tests to fail and the other succeed. So far I have not found one that causes both to succeed.