Change evaluate::function return types to Evaluated (#966)
change return type Result<Value> to Result<Evaluated>

All existing functions operating in evaluate are evaluated by enum Value without having to have an intermediate value, Evaluated. However, if it is used in a future query builder, if it is a chained structure in a function such as a substring that should always hold an intermediate value, an overhead may occur by repeating the clone.
In this case, it can be solved by string slicing to keep the intermediate value. Modify the return values ​​for this structure.