Expose wasm-pack's init function argument to users (#971)
Some users need to be able to use the `module_or_path` argument of the function `init` that `wasm-pack` generates. In my case, it enables loading the WASM module in an environment that is neither the browser nor node (edge).
This modification can fix WASM loading issues for other users too.