Add enum `StrSlice` under enum `Evaluated` (#999)
Description
If a function that handles a string is chained and executed on the query builder, the string is copied every time by the Clone trait. This overhead is solved by always having a sliced ​​enum with a value.

Implementation plan
#[derive(Clone)]
pub enum Evaluated<'a> {
    Literal(Literal<'a>),
    StrSlice { source: String, range: Range<usize> },
    Value(Value),
}
If a variant called source holds the source of the string, and the string function is called continuously after that, the range is updated every time.