Add `CALC_DISTANCE` function (#1153)
CALC_DISTANCE

calc_distance is a function used in spatial analysis to calculate the distance between two geometries. It is commonly used in geographic information systems and spatial databases to determine the proximity of objects and perform various spatial analyses. The function takes two geometries as inputs and returns the shortest distance between them. It can operate on several geometry types, including points, lines, polygons, and multi-geometries. The result of calc_distance is expressed in the same unit of measurement as the input geometries, making it easy to compare distances between different objects.